Online-Shop/requirements.txt
Django==4.2.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-spectacular==0.26.5
celery==5.3.6
redis==5.0.1
Pillow==10.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
